数据库查询数据

+++ src/main/java/zju/bpamodel/swi/PV.java
@@ -22,8 +22,8 @@ public class PV implements Serializable {
     private double isc;
     private double um;
     private double im;
-    private double n1;
-    private double n2;
+    private int n1;
+    private int n2;
 
     public static PV createPV(String content) {
         PV pv = new PV();
@@ -155,19 +155,19 @@ public class PV implements Serializable {
         this.im = im;
     }
 
-    public double getN1() {
+    public int getN1() {
         return n1;
     }
 
-    public void setN1(double n1) {
+    public void setN1(int n1) {
         this.n1 = n1;
     }
 
-    public double getN2() {
+    public int getN2() {
         return n2;
     }
 
-    public void setN2(double n2) {
+    public void setN2(int n2) {
         this.n2 = n2;
     }
 }
